perf(ui): split task and data quality utilities
