+++ openmetadata-ui/src/main/resources/ui/src/components/DataContract/ContractSemanticFormTab/ContractSemanticFormTab.tsx
@@ -13,6 +13,7 @@
 
 import Icon from '@ant-design/icons';
 import { Actions, JsonTree } from '@react-awesome-query-builder/antd';
+import '@react-awesome-query-builder/antd/css/styles.css';
 import {
   Button,
   Col,
