+++ openmetadata-ui/src/main/resources/ui/src/pages/TasksPage/shared/DescriptionTask.tsx
@@ -21,7 +21,7 @@ import {
   Thread,
   ThreadTaskStatus,
 } from '../../../generated/entity/feed/thread';
-import { getDescriptionDiff } from '../../../utils/TasksUtils';
+import { getDescriptionDiff } from '../../../utils/TaskPayloadUtils';
 import { DescriptionTabs } from './DescriptionTabs';
 import DiffView from './DiffView/DiffView';
